refactor: Clean up tests.
- Move helpers into their own file.
- Use named properties for struct literals.
