+++ builders/state_test.go
@@ -6,41 +6,14 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/testdouble/diplomat/builders"
 	"github.com/testdouble/diplomat/errors"
-	"github.com/testdouble/diplomat/http"
 	"github.com/testdouble/diplomat/parsers"
 )
-
-func assertTest(t *testing.T, expected builders.Test, actual builders.Test, err error) {
-	if err != nil {
-		assert.FailNow(t, "Error should not exist.")
-		return
-	}
-	assert.Equal(t, expected, actual)
-}
-
-func fillRequest(method string, path string, headers map[string]string, body string) *http.Request {
-	request := http.NewRequest(method, path)
-	for key, value := range headers {
-		request.Headers[key] = value
-	}
-	request.Body = []byte(body)
-	return request
-}
-
-func fillResponse(code int, status string, headers map[string]string, body string) *http.Response {
-	response := http.NewResponse(code, status)
-	for key, value := range headers {
-		response.Headers[key] = value
-	}
-	response.Body = []byte(body)
-	return response
-}
 
 func TestNoBody(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"< PROTO 1337 STATUS TEXT",
@@ -64,8 +37,8 @@ func TestNoBody(t *testing.T) {
 func TestSingleLineBody(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"< PROTO 1337 STATUS TEXT",
@@ -91,8 +64,8 @@ func TestSingleLineBody(t *testing.T) {
 func TestMultiLineBodyWithIndentation(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"< PROTO 1337 STATUS TEXT",
@@ -119,8 +92,8 @@ func TestMultiLineBodyWithIndentation(t *testing.T) {
 func TestMissingBracket(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"< PROTO 1337 STATUS TEXT",
@@ -145,8 +118,8 @@ func TestMissingBracket(t *testing.T) {
 func TestMultiLineBodyWithAngleBracket(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"< PROTO 1337 STATUS TEXT",
@@ -166,8 +139,8 @@ func TestMultiLineBodyWithAngleBracket(t *testing.T) {
 func TestRequestBody(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 			"Some request body",
@@ -194,8 +167,8 @@ func TestRequestBody(t *testing.T) {
 func TestRequestOnly(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"> METHOD path",
 			"> Header: Request",
 		},
@@ -209,8 +182,8 @@ func TestRequestOnly(t *testing.T) {
 func TestResponseOnly(t *testing.T) {
 	subject := builders.State{}
 	body := parsers.Spec{
-		"",
-		[]string{
+		Name: "",
+		Body: []string{
 			"< HTTP/1.1 200 OK",
 			"< Content-Length: 0",
 			"<",
